requirements.txt
transformers
transformers[sentencepiece]
datasets
torch
matplotlib
pandas
numpy
sacrebleu
rouge_score
py7zr
nltk
tqdm
pyYAML
notebook
boto3
mypy-boto3-s3
python-box==6.0.2
ensure-annotations
fastapi==0.78.0
uvicorn==0.18.3
Jinja2==3.1.2
-e .

